% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_etm.R
\name{gen_data_etm}
\alias{gen_data_etm}
\title{Generate data in a stepped wedge design with a exposure time treatment effect}
\usage{
gen_data_etm(
  n,
  J,
  m,
  beta,
  gamma,
  betas,
  betaM,
  gammas,
  sigma_a,
  sigma_ey,
  sigma_tau,
  sigma_em,
  binary.outcome = 0,
  binary.mediator = 0
)
}
\arguments{
\item{n}{the number of cluster}

\item{J}{the number of time periods}

\item{m}{the number of individuals per each cluster}

\item{beta}{the underling time trend effect in a outcome model}

\item{gamma}{the underling time trend effect in a mediator model}

\item{betas}{the exposure effect (a J-1 vector) on the outcome conditional on mediaotr and other possible confounders}

\item{betaM}{the mediator effect on the outcome conditional on other possible confounders}

\item{gammas}{the exposure effect (a J-1 vector) on the mediator conditional other possible confounders}

\item{sigma_a}{the standard error of random effect in outcome model (assuming it follows normal distribution with mean zero)}

\item{sigma_ey}{the standard error of error term in outcome model (assuming it follows normal distribution with mean zero)}

\item{sigma_tau}{the standard error of random effect in mediator model (assuming it follows normal distribution with mean zero)}

\item{sigma_em}{the standard error of error term in mediator model (assuming it follows normal distribution with mean zero)}

\item{binary.outcome}{(Required) If the outcome is binary, set to 1. If the outcome is continuous, set to 0}

\item{binary.mediator}{description (Required) If the mediator is binary, set to 1. If the mediator is continuous, set to 0}
}
\value{
A data frame including cluster, period, id, S (time since intervention), A (treatment state indicator),
c (the start time of the treatment), alpha (random effects of observations within a cluster), tau (random effects of mediator within a cluster),
M (mediator), Y (outcome)
}
\description{
This function can generate longitudinal data in a stepped wedge design in the presence of
a exposure time treatment effect
}
\examples{
n = 9  
J = 4
m = 20
beta = cumsum(c(0,0.1,0.1/2,0.1/(2^2)))
gamma = cumsum(c(0,0.3,0.3/2,0.3/(2^2)))
sigma_tau = sigma_a = 0.334
sigma_em = sigma_ey = 1
gammas = c(0.32,0.4,0.48) 
betas = c(0.6,0.75,0.9)
# generate continuous outcome and continuous mediator
mydata1 = gen_data_etm(n,J,m,beta,gamma,betas,betaM=0.625,gammas,sigma_a,sigma_ey,
sigma_tau,sigma_em,binary.outcome=0,binary.mediator=0)

# generate continuous outcome and binary mediator
sigma_tau = 0.605    #using this value can make sure ICC of M is about 0.1 since  
#ICC of binary variable is computed as sigma_tau^2/(sigma_tau^2+pi^2/3)
mydata2 = gen_data_etm(n,J,m,beta,gamma,betas,betaM=0.625,gammas,sigma_a,sigma_ey,
sigma_tau,sigma_em,binary.outcome=0,binary.mediator=1)

# generate binary outcome and continuous mediator
sigma_a = 0.605
mydata3 = gen_data_etm(n,J,m,beta,gamma,betas,betaM=0.625,gammas,sigma_a,sigma_ey,
sigma_tau,sigma_em,binary.outcome=1,binary.mediator=0)

# generate binary outcome and continuous mediator
sigma_tau = sigma_a = 0.605
mydata4 = gen_data_hhm(n,J,m,beta,gamma,betaA=0.75,betaM=0.625,gammaA=0.4,sigma_a,
sigma_ey,sigma_tau,sigma_em,binary.outcome=1,binary.mediator=1)

# generate continuous outcome and continuous mediator with different clusters and period
n = 12  
J = 5
beta = cumsum(c(0,0.1,0.1/2,0.1/(2^2),0.1/(2^3)))
gamma = cumsum(c(0,0.3,0.3/2,0.3/(2^2),0.3/(2^3)))
gammas = c(0.32,0.4,0.48,0.56) 
betas = c(0.6,0.75,0.9,1.05)
sigma_tau = sigma_a = 0.334
sigma_em = sigma_ey = 1
mydata5 = gen_data_etm(n,J,m,beta,gamma,betas,betaM=0.625,gammas,sigma_a,sigma_ey,
sigma_tau,sigma_em,binary.outcome=0,binary.mediator=0)
}
\references{
Kenny A., Voldal E.C., Xia F., Heagerty P.J. and Hughes J.P. Analysis of stepped wedge cluster randomized trials in the
presence of a time-varying treatment effect. Statistics in Medicine. 2022;41(22):4311-4339.
}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk} and Fan Li \email{fan.f.li@yale.edu}
}
