% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_hhm.R
\name{gen_data_hhm}
\alias{gen_data_hhm}
\title{Generate data in a stepped wedge design based on Hussey and Hughes model with a constant treatment effect}
\usage{
gen_data_hhm(
  I,
  J,
  n,
  beta,
  gamma,
  theta,
  beta_M,
  eta,
  sigma_a,
  sigma_ey,
  sigma_tau,
  sigma_em,
  binary.o = 0,
  binary.m = 0
)
}
\arguments{
\item{I}{The number of cluster}

\item{J}{The number of time periods}

\item{n}{The number of individuals per each cluster}

\item{beta}{The underling time trend effect in a outcome model}

\item{gamma}{The underling time trend effect in a mediator model}

\item{theta}{The exposure effect on the outcome conditional on mediator and other possible confounders}

\item{beta_M}{The mediator effect on the outcome conditional on other possible confounders}

\item{eta}{The exposure effect on the mediator conditional on other possible confounders}

\item{sigma_a}{The standard error of random effect in outcome model (assuming it follows normal distribution with mean zero)}

\item{sigma_ey}{The standard error of error term in outcome model (assuming it follows normal distribution with mean zero)}

\item{sigma_tau}{The standard error of random effect in mediator model (assuming it follows normal distribution with mean zero)}

\item{sigma_em}{The standard error of error term in mediator model (assuming it follows normal distribution with mean zero)}

\item{binary.o}{(Required) If the outcome is binary, set to 1. If the outcome is continuous, set to 0}

\item{binary.m}{(Required) If the mediator is binary, set to 1. If the mediator is continuous, set to 0}
}
\value{
A data frame including cluster, period, id, E (time since intervention, i.e., exposure time), A (treatment status indicator),
c (the start time of the treatment), alpha (random effects of observations within a cluster), tau (random effects of mediator within a cluster),
M (mediator), Y (outcome)
}
\description{
This function can generate longitudinal data in a stepped wedge design based on Hussey and Hughes model
with a constant treatment effect.
}
\examples{
I = 9  
J = 4
n = 20
beta = cumsum(c(0,0.1,0.1/2,0.1/(2^2)))
gamma = cumsum(c(0,0.3,0.3/2,0.3/(2^2)))
sigma_tau = sigma_a = 0.334
sigma_em = sigma_ey = 1
# generate continuous outcome and continuous mediator
mydata1 = gen_data_hhm(I,J,n,beta,gamma,theta=0.75,beta_M=0.625,eta=0.4,sigma_a,
sigma_ey,sigma_tau,sigma_em,binary.o=0,binary.m=0)

# generate continuous outcome and binary mediator
sigma_tau = 0.605    #using this value can produce ICC of mediator model is about 0.1 since  
# ICC of binary variable is calculated as sigma_tau^2/(sigma_tau^2+pi^2/3)
mydata2 = gen_data_hhm(I,J,n,beta,gamma,theta=0.75,beta_M=0.625,eta=0.4,sigma_a,
sigma_ey,sigma_tau,sigma_em,binary.o=0,binary.m=1)

# generate binary outcome and continuous mediator
sigma_a = 0.605
mydata3 = gen_data_hhm(I,J,n,beta,gamma,theta=0.75,beta_M=0.625,eta=0.4,sigma_a,
sigma_ey,sigma_tau,sigma_em,binary.o=1,binary.m=0)

# generate binary outcome and binary mediator
sigma_tau = sigma_a = 0.605
mydata4 = gen_data_hhm(I,J,n,beta,gamma,theta=0.75,beta_M=0.625,eta=0.4,sigma_a,
sigma_ey,sigma_tau,sigma_em,binary.o=1,binary.m=1)

# generate continuous outcome and continuous mediator with different clusters and period
I = 12  
J = 5
beta = cumsum(c(0,0.1,0.1/2,0.1/(2^2),0.1/(2^3)))
gamma = cumsum(c(0,0.3,0.3/2,0.3/(2^2),0.3/(2^3)))
sigma_tau = sigma_a = 0.334
sigma_em = sigma_ey = 1
mydata5 = gen_data_hhm(I,J,n,beta,gamma,theta=0.75,beta_M=0.625,eta=0.4,sigma_a,
sigma_ey,sigma_tau,sigma_em,binary.o=0,binary.m=0)
}
\references{
Hussey M.A. and Hughes J.P. Design and analysis of stepped wedge cluster randomized trials. Contemporary Clinical Trials.
2007;28(2):182-191.

Kenny A., Voldal E.C., Xia F., Heagerty P.J. and Hughes J.P. Analysis of stepped wedge cluster randomized trials in the
presence of a time-varying treatment effect. Statistics in Medicine. 2022;41(22):4311-4339.
}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk} and Fan Li \email{fan.f.li@yale.edu}
}
