% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_data_nem_etm.R
\name{gen_data_nem_etm}
\alias{gen_data_nem_etm}
\title{Generate data in a stepped wedge design based on nested exchangeable correlation model with an exposure-time dependent treatment effect}
\usage{
gen_data_nem_etm(
  I,
  J,
  n,
  beta,
  gamma,
  theta_e,
  beta_M,
  eta_e,
  sigma_a,
  sigma_phi,
  sigma_ey,
  sigma_tau,
  sigma_psi,
  sigma_em,
  binary.o = 0,
  binary.m = 0
)
}
\arguments{
\item{I}{The number of cluster}

\item{J}{The number of time periods}

\item{n}{The number of individuals per each cluster}

\item{beta}{The underling time trend effect in a outcome model}

\item{gamma}{The underling time trend effect in a mediator model}

\item{theta_e}{The exposure effect (a J-1 vector) on the outcome conditional on mediator and other possible confounders}

\item{beta_M}{The mediator effect on the outcome conditional on other possible confounders}

\item{eta_e}{The exposure effect (a J-1 vector) on the mediator conditional on other possible confounders}

\item{sigma_a}{The standard error of random effect in outcome model (assuming it follows normal distribution with mean zero)}

\item{sigma_phi}{The standard error of random effect of cluster-by-time interaction in outcome model (assuming it follows normal distribution with mean zero)}

\item{sigma_ey}{The standard error of error term in outcome model (assuming it follows normal distribution with mean zero)}

\item{sigma_tau}{The standard error of random effect in mediator model (assuming it follows normal distribution with mean zero)}

\item{sigma_psi}{The standard error of random effect of cluster-by-time interaction in mediator model (assuming it follows normal distribution with mean zero)}

\item{sigma_em}{The standard error of error term in mediator model (assuming it follows normal distribution with mean zero)}

\item{binary.o}{(Required) If the outcome is binary, set to 1. If the outcome is continuous, set to 0}

\item{binary.m}{(Required) If the mediator is binary, set to 1. If the mediator is continuous, set to 0}
}
\value{
A data frame including cluster, period, id, E (time since intervention, i.e., exposure time), A (treatment status indicator),
c (the start time of the treatment), alpha (random effects of outcome within a cluster), phi (random effects of cluster-by-time interaction in outcome model),
tau (random effects of mediator within a cluster), psi (random effects of cluster-by-time interaction in mediator model), M (mediator), Y (outcome)
}
\description{
This function can generate longitudinal data in a stepped wedge design nested exchangeable correlation model in the presence of
an exposure-time dependent treatment effect
}
\examples{
I = 9  
J = 4
n = 20
beta = cumsum(c(0,0.1,0.1/2,0.1/(2^2)))
gamma = cumsum(c(0,0.3,0.3/2,0.3/(2^2)))
sigma_tau = sigma_a = 0.334
sigma_phi = sigma_psi = 0.6
sigma_em = sigma_ey = 0.8
eta_e = c(0.32,0.4,0.48) 
theta_e = c(0.6,0.75,0.9)
# generate continuous outcome and continuous mediator
mydata1 = gen_data_nem_etm(I,J,n,beta,gamma,theta_e,beta_M=0.625,eta_e,sigma_a,
sigma_phi,sigma_ey,sigma_tau,sigma_psi,sigma_em,binary.o=0,binary.m=0)

# generate continuous outcome and binary mediator
mydata2 = gen_data_nem_etm(I,J,n,beta,gamma,theta_e,beta_M=0.625,eta_e,sigma_a,
sigma_phi,sigma_ey,sigma_tau,sigma_psi,sigma_em,binary.o=0,binary.m=1)

# generate binary outcome and continuous mediator
mydata3 = gen_data_nem_etm(I,J,n,beta,gamma,theta_e,beta_M=0.625,eta_e,sigma_a,
sigma_phi,sigma_ey,sigma_tau,sigma_psi,sigma_em,binary.o=1,binary.m=0)

# generate binary outcome and binary mediator
mydata4 = gen_data_nem_etm(I,J,n,beta,gamma,theta_e,beta_M=0.625,eta_e,sigma_a,
sigma_phi,sigma_ey,sigma_tau,sigma_psi,sigma_em,binary.o=1,binary.m=1)

# generate continuous outcome and continuous mediator with different clusters and period
I = 12  
J = 5
beta = cumsum(c(0,0.1,0.1/2,0.1/(2^2),0.1/(2^3)))
gamma = cumsum(c(0,0.3,0.3/2,0.3/(2^2),0.3/(2^3)))
eta_e = c(0.32,0.4,0.48,0.56) 
theta_e = c(0.6,0.75,0.9,1.05)
mydata5 = gen_data_nem_etm(I,J,n,beta,gamma,theta_e,beta_M=0.625,eta_e,sigma_a,
sigma_phi,sigma_ey,sigma_tau,sigma_psi,sigma_em,binary.o=0,binary.m=0)

}
\references{
Li F., Hughes J.P., Hemming K., Taljaard M., Melnick E.R. and Heagerty P.J. Mixed-effects models for the design
and analysis of stepped wedge cluster randomized trials: An overview. Statistical Methods in Medical Research. 2021;30(2):612-639.

Kenny A., Voldal E.C., Xia F., Heagerty P.J. and Hughes J.P. Analysis of stepped wedge cluster randomized trials in the
presence of a time-varying treatment effect. Statistics in Medicine. 2022;41(22):4311-4339.
}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk} and Fan Li \email{fan.f.li@yale.edu}
}
