% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate_contY_binaM_hhm.R
\name{mediate_contY_binaM_hhm}
\alias{mediate_contY_binaM_hhm}
\title{Mediation analysis function for continuous outcome and binary mediator in a stepped wedge design
based on Hussey and Hughes model with a constant treatment effect}
\usage{
mediate_contY_binaM_hhm(
  data,
  method = "STA",
  outcome = "Y",
  mediator = "M",
  treatment = "A",
  cluster = "cluster",
  period = "period",
  exposure = "E",
  covariateY = NULL,
  covariateM = NULL,
  a0 = 0,
  a1 = 1
)
}
\arguments{
\item{data}{A dataset, which should include outcome, mediator, treatment, cluster, period variables}

\item{method}{Two approximate methods are provided, one is STA (second-order Taylor approximate), the other is GHQ (Gauss-Hermite Quadrature)}

\item{outcome}{The outcome variable in stepped wedge design}

\item{mediator}{The mediator variable in stepped wedge design}

\item{treatment}{The treatment variable in stepped wedge design. When there is a implementation period, then corresponding treatment status
should be set to -1, which is mainly for convenience}

\item{cluster}{The cluster variable in stepped wedge design}

\item{period}{The period variable in stepped wedge design}

\item{exposure}{The exposure time with J levels (e.g., 0,1,2,...,J-1 when there is no implementation period). If dataset has no exposure, then set it to NULL}

\item{covariateY}{A vector of confounders in the outcome regression (default is NULL)}

\item{covariateM}{A vector of confounders in the mediator regression  (default is NULL)}

\item{a0}{The reference treatment level in defining TE and NIE. The default value is 0}

\item{a1}{The treatment level of interest in defining TE and NIE. The default value is 1}
}
\value{
A list containing point and interval estimates of NIE, NDE, TE and MP.
}
\description{
After obtaining parameter estimates from linear mixed effect model for outcome model and
generalized linear mixed effect model for mediator model, this function can obtain
mediation measures including NIE, NDE, TE and MP assuming a constant treatment effect
}
\examples{
library(lme4)
I = 15 
J = 4
n = 20
beta = cumsum(c(0,0.1,0.1/2,0.1/(2^2)))
gamma = cumsum(c(0,0.3,0.3/2,0.3/(2^2)))
sigma_a = 0.334
sigma_tau = 0.605   
sigma_em = sigma_ey = 1
set.seed(123456)
mydata1 = gen_data_hhm(I,J,n,beta,gamma,theta=0.75,beta_M=1.2,eta=0.6,
sigma_a,sigma_ey,sigma_tau,sigma_em,binary.o=0,binary.m=1)
# example 1: mediation analysis without covariates in outcome and mediator models 
res1 = mediate_contY_binaM_hhm(data=mydata1, method = "STA")
print(res1)
res1f = mediate_contY_binaM_hhm(data=mydata1, method = "GHQ")
print(res1f)

# example 2: mediation analysis with different covariates in outcome and mediator models
# generate two covariates
covdata = data.frame("X1" = numeric(), "X2" = numeric())
set.seed(100)
for(i in 1:I){
  for(j in 1:J){
     x1_ijk = rnorm(n,mean=0.5) 
     x2_ijk = rnorm(n,sd=0.5)
     covdata = rbind(covdata, data.frame(cbind(X1 = x1_ijk, X2 = x2_ijk)))
  }
}

mydata2 = data.frame(mydata1,covdata)
res2 = mediate_contY_binaM_hhm(data=mydata2, method = "STA",
covariateY = c("X1"), covariateM = c("X2"))
print(res2)
res2f = mediate_contY_binaM_hhm(data=mydata2, method = "GHQ",
covariateY = c("X1"), covariateM = c("X2"))
print(res2f)

# example 3: mediation analysis with the same covariates in outcome and mediator models
res3 = mediate_contY_binaM_hhm(data=mydata2, method = "STA",
covariateY = c("X1","X2"), covariateM = c("X1","X2"))
print(res3)
res3f = mediate_contY_binaM_hhm(data=mydata2, method = "GHQ",
covariateY = c("X1","X2"), covariateM = c("X1","X2"))
print(res3f)

# example 4: mediation analysis with different clusters and periods
I = 12  
J = 5
beta = cumsum(c(0,0.1,0.1/2,0.1/(2^2),0.1/(2^3)))
gamma = cumsum(c(0,0.3,0.3/2,0.3/(2^2),0.3/(2^3)))
set.seed(123456)
mydata3 = gen_data_hhm(I,J,n,beta,gamma,theta=0.75,beta_M=1,eta=0.8,
sigma_a,sigma_ey,sigma_tau,sigma_em,binary.o=0,binary.m=1)
res4 = mediate_contY_binaM_hhm(data=mydata3,method = "STA")
print(res4)
res4f = mediate_contY_binaM_hhm(data=mydata3,method = "GHQ")
print(res4f)

# example 5: if there are no exposure variable
mydata4 = mydata3[,-4]
res5 = mediate_contY_binaM_hhm(data=mydata4,method = "STA",exposure=NULL)
print(res5)
}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk} and Fan Li \email{fan.f.li@yale.edu}
}
