% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediate_contY_contM_etm.R
\name{mediate_contY_contM_etm}
\alias{mediate_contY_contM_etm}
\title{Mediation analysis function for continuous outcome and continuous mediator in a stepped wedge design with an exposure-time dependent treatment effect}
\usage{
mediate_contY_contM_etm(
  data,
  outcome = "Y",
  mediator = "M",
  treatment = "A",
  cluster = "cluster",
  period = "period",
  exposure = "E",
  covariateY = NULL,
  covariateM = NULL,
  a0 = 0,
  a1 = 1
)
}
\arguments{
\item{data}{A dataset, which should include outcome, mediator, treatment, cluster, period variables}

\item{outcome}{The outcome variable in stepped wedge design}

\item{mediator}{The mediator variable in stepped wedge design}

\item{treatment}{The treatment variable in stepped wedge design. When there is a implementation period, then corresponding treatment status
should be set to -1, which is mainly for convenience}

\item{cluster}{The cluster variable in stepped wedge design}

\item{period}{The period variable in stepped wedge design}

\item{exposure}{The exposure time with J levels (e.g., 0,1,2,...,J-1 when there is no implementation period)}

\item{covariateY}{A vector of confounders in the outcome regression (default is NULL)}

\item{covariateM}{A vector of confounders in the mediator regression  (default is NULL)}

\item{a0}{The reference treatment level in defining TE and NIE. The default value is 0}

\item{a1}{The treatment level of interest in defining TE and NIE. The default value is 1}
}
\value{
A list containing point and interval estimates of NIE, NDE, TE and MP at each exposure time and
overall summary mediation measures, as well as Chi-square test result of TE
}
\description{
After obtaining parameter estimates from linear mixed-effects models for both outcome and mediator models, this function can obtain
mediation measures including NIE, NDE, TE and MP with an exposure-time dependent treatment effect
}
\examples{
library(lme4)
I = 15 
J = 4
n = 20
beta = cumsum(c(0,0.1,0.1/2,0.1/(2^2)))
gamma = cumsum(c(0,0.3,0.3/2,0.3/(2^2)))
eta_e = c(0.5,0.8,1.1) 
theta_e = c(0.5,1.0,1.5)
sigma_tau = sigma_a = 0.334
sigma_em = sigma_ey = 1
set.seed(123456)
mydata1 = gen_data_etm(I,J,n,beta,gamma,theta_e,beta_M=0.8,eta_e,sigma_a,
sigma_ey,sigma_tau,sigma_em,binary.o=0,binary.m=0)
# example 1: mediation analysis without covariates in outcome and mediator models
res1 = mediate_contY_contM_etm(data = mydata1)
print(res1)

# example 2: mediation analysis with different covariates in outcome and mediator models
# generate two covariates
covdata = data.frame("X1" = numeric(), "X2" = numeric())
set.seed(100)
for(i in 1:I){
  for(j in 1:J){
     x1_ijk = rnorm(n,mean=0.5) 
     x2_ijk = rnorm(n,sd=0.5)
     covdata = rbind(covdata, data.frame(cbind(X1 = x1_ijk, X2 = x2_ijk)))
  }
}
mydata2 = data.frame(mydata1,covdata)
res2 = mediate_contY_contM_etm(data = mydata2, covariateY = c("X1"), 
covariateM = c("X2"))
print(res2)

# example 3: mediation analysis with the same covariates in outcome and mediator models
res3 = mediate_contY_contM_etm(data = mydata2, covariateY = c("X1","X2"), 
covariateM = c("X1","X2"))
print(res3)

# example 4: mediation analysis with different clusters and periods
I = 12  
J = 5
beta = cumsum(c(0,0.1,0.1/2,0.1/(2^2),0.1/(2^3)))
gamma = cumsum(c(0,0.3,0.3/2,0.3/(2^2),0.3/(2^3)))
eta_e = c(0.4,0.6,0.8,1.0) 
theta_e = c(0.6,0.75,0.9,1.05)
set.seed(123456)
mydata3 = gen_data_etm(I,J,n,beta,gamma,theta_e,beta_M=0.8,eta_e,sigma_a,sigma_ey,
sigma_tau,sigma_em,binary.o=0,binary.m=0)
res4 = mediate_contY_contM_etm(data=mydata3)
print(res4)

}
\author{
Zhiqiang Cao \email{zcaoae@connect.ust.hk} and Fan Li \email{fan.f.li@yale.edu}
}
